✨ Add hero context menu

+++ src/routes/tabletop/$campaignId.gm/-components/SettingsPanel/UnitsTab/Heroes.tsx
@@ -1,15 +1,17 @@
 import { ActionIcon, Avatar, Button, Group, Menu, Modal, NumberInput, Stack, Table, Text, Title, type NumberInputProps } from '@mantine/core'
 import { useForm } from '@mantine/form'
 import { useDisclosure } from '@mantine/hooks'
+import { notifications } from '@mantine/notifications'
 import { IconPencil, IconTrash, IconX } from '@tabler/icons-react'
 import { useMutation } from '@tanstack/react-query'
+import { getRouteApi } from '@tanstack/react-router'
 import { createServerFn } from '@tanstack/react-start'
 import { type } from 'arktype'
 import { Fragment } from 'react'
-import { type Tables } from '~/supabase/databaseTypes'
-import { getSupabaseServerClient } from '~/supabase/getSupabaseServerClient'
+import { getServiceClient } from '~/supabase/getServiceClient'
+import { requireAccount } from '~/supabase/requireAccount'
 import { useSupabase } from '~/supabase/useSupabase'
-import { useTabletopHeroes } from '../../../-hooks/-useTabletopData'
+import { useTabletopHeroes } from '../../../-hooks/useTabletopData'
 
 export default function Heroes() {
 	const { data: heroes } = useTabletopHeroes()
@@ -48,32 +50,42 @@ type HeroProps = {
 }
 
 function Hero({ heroId }: HeroProps) {
+	const { campaignId } = getRouteApi('/tabletop/$campaignId/gm/').useLoaderData()
 	const { heroName, tabletopHero } = useTabletopHeroes().data[heroId] ?? {}
 
 	const [editMode, editModeHandlers] = useDisclosure(false)
 
 	const removeHero = useMutation({
-		mutationFn: removeHeroAction
+		mutationFn: removeHeroAction,
+		onError: error => {
+			notifications.show({
+				title: 'Failed to remove hero',
+				color: 'red',
+				message: error.message
+			})
+		}
 	})
 
-	const getTabletopLifeValue = (value: keyof Tables<'tabletop_heroes'> & (`health_${string}` | `shield_${string}`)) => {
-		return tabletopHero?.[value] ?? 0
-	}
+	// const getTabletopLifeValue = (value: keyof Tables<'tabletop_heroes'> & (`health_${string}` | `shield_${string}`)) => {
+	// 	return tabletopHero?.[value] ?? 0
+	// }
 
-	const getTabletopPosition = () => {
-		if (!tabletopHero) return null
-		if (
-			tabletopHero.position_q === null
-			|| tabletopHero.position_r === null
-			|| tabletopHero.position_s === null
-		) return null
-		const position = {
-			q: tabletopHero.position_q,
-			r: tabletopHero.position_r,
-			s: tabletopHero.position_s
-		}
-		return `${position.q},${position.r},${position.s}`
-	}
+	// const getTabletopPosition = () => {
+	// if (!tabletopHero) return null
+	// if (
+	// 	tabletopHero.position_q === null
+	// 	|| tabletopHero.position_r === null
+	// 	|| tabletopHero.position_s === null
+	// ) return null
+	// const position = {
+	// 	q: tabletopHero.position_q,
+	// 	r: tabletopHero.position_r,
+	// 	s: tabletopHero.position_s
+	// }
+	// return `${position.q},${position.r},${position.s}`
+	// }
+
+	if (!tabletopHero) return null
 
 	return (
 		<Fragment>
@@ -92,12 +104,12 @@ function Hero({ heroId }: HeroProps) {
 				</Table.Td>
 				<Table.Td>
 					<Stack gap={0}>
-						<Text size='sm'>{getTabletopLifeValue('shield_durability')}</Text>
-						<Text size='sm'>{getTabletopLifeValue('shield_current')} / {getTabletopLifeValue('shield_max')}</Text>
+						{/* <Text size='sm'>{getTabletopLifeValue('shield_durability')}</Text>
+						<Text size='sm'>{getTabletopLifeValue('shield_current')} / {getTabletopLifeValue('shield_max')}</Text> */}
 					</Stack>
 				</Table.Td>
-				<Table.Td>{getTabletopLifeValue('health_current')} / {getTabletopLifeValue('health_max')}</Table.Td>
-				<Table.Td>{getTabletopPosition() ?? 'Unplaced'}</Table.Td>
+				{/* <Table.Td>{getTabletopLifeValue('health_current')} / {getTabletopLifeValue('health_max')}</Table.Td> */}
+				{/* <Table.Td>{getTabletopPosition() ?? 'Unplaced'}</Table.Td> */}
 				<Table.Td>
 					<Group>
 						<ActionIcon variant='subtle' color='gray' onClick={editModeHandlers.toggle}>
@@ -114,7 +126,12 @@ function Hero({ heroId }: HeroProps) {
 								<Menu.Item
 									color='red'
 									leftSection={<IconTrash size={14} />}
-									onClick={() => removeHero.mutate({ data: { hero_id: heroId } })}
+									onClick={() => removeHero.mutate({
+										data: {
+											campaignId,
+											characterId: tabletopHero.characterId
+										}
+									})}
 								>
 									Remove Hero
 								</Menu.Item>
@@ -128,18 +145,31 @@ function Hero({ heroId }: HeroProps) {
 }
 
 const removeHeroSchema = type({
-	hero_id: 'number'
+	characterId: 'number',
+	campaignId: 'number'
 })
 
 const removeHeroAction = createServerFn({ method: 'POST' })
 	.validator(removeHeroSchema)
-	.handler(async ({ data: { hero_id } }) => {
-		const supabase = getSupabaseServerClient()
+	.handler(async ({ data: { campaignId, characterId } }) => {
+		const { supabase, user } = await requireAccount()
 
-		const { error } = await supabase
-			.from('tabletop_heroes')
+		// Check if the user is the GM
+		const gmCheck = await supabase
+			.from('campaign_info')
+			.select('gmUserId: user_id')
+			.eq('campaign_id', campaignId)
+			.limit(1)
+			.single()
+		if (gmCheck.error) throw new Error(gmCheck.error.message, { cause: gmCheck.error })
+		if (gmCheck.data.gmUserId !== user.id) throw new Error('You are not the GM of this campaign')
+
+		const serviceClient = getServiceClient()
+
+		const { error } = await serviceClient
+			.from('tabletop_characters')
 			.delete()
-			.eq('hero_id', hero_id)
+			.eq('character_id', characterId)
 		if (error) throw new Error(error.message, { cause: error })
 	})
 
@@ -164,33 +194,33 @@ function HeroEditModal({ heroId, opened, close }: HeroEditModalProps) {
 	const form = useForm({
 		mode: 'uncontrolled',
 		initialValues: {
-			shieldDurability: tabletopHero?.shield_durability ?? 0,
-			shieldCurrent: tabletopHero?.shield_current ?? 0,
-			shieldMax: tabletopHero?.shield_max ?? 0,
-			healthCurrent: tabletopHero?.health_current ?? 0,
-			healthMax: tabletopHero?.health_max ?? 0
+			// shieldDurability: tabletopHero?.shield_durability ?? 0,
+			// shieldCurrent: tabletopHero?.shield_current ?? 0,
+			// shieldMax: tabletopHero?.shield_max ?? 0,
+			// healthCurrent: tabletopHero?.health_current ?? 0,
+			// healthMax: tabletopHero?.health_max ?? 0
 		},
 		validate: {
-			healthMax: value => value > 0 ? null : 'Max health must be greater than 0'
+			// healthMax: value => value > 0 ? null : 'Max health must be greater than 0'
 		}
 	})
 
-	const handleSubmit = (values: typeof form.values) => {
-		(async () => {
-			const { error } = await supabase
-				.from('tabletop_heroes')
-				.upsert({
-					hero_id: heroId,
-					shield_durability: values.shieldDurability,
-					shield_current: values.shieldCurrent,
-					shield_max: values.shieldMax,
-					health_current: values.healthCurrent,
-					health_max: values.healthMax
-				})
-			if (error) throw new Error(error.message, { cause: error })
-		})().catch(console.error)
-		close()
-	}
+	// const handleSubmit = (values: typeof form.values) => {
+	// 	(async () => {
+	// 		const { error } = await supabase
+	// 			.from('tabletop_heroes')
+	// 			.upsert({
+	// 				hero_id: heroId,
+	// 				shield_durability: values.shieldDurability,
+	// 				shield_current: values.shieldCurrent,
+	// 				shield_max: values.shieldMax,
+	// 				health_current: values.healthCurrent,
+	// 				health_max: values.healthMax
+	// 			})
+	// 		if (error) throw new Error(error.message, { cause: error })
+	// 	})().catch(console.error)
+	// 	close()
+	// }
 
 	return (
 		<Modal
@@ -199,48 +229,48 @@ function HeroEditModal({ heroId, opened, close }: HeroEditModalProps) {
 			onExitTransitionEnd={() => form.reset()}
 			title={`Edit ${heroName}`}
 		>
-			<form onSubmit={form.onSubmit(handleSubmit)}>
-				<Stack>
-					<Group grow>
-						<NumberInput
-							label='Shield Durability'
-							key={form.key('shieldDurability')}
-							{...form.getInputProps('shieldDurability')}
-							{...numberInputProps}
-						/>
-						<NumberInput
-							label='Current Shield'
-							key={form.key('shieldCurrent')}
-							{...form.getInputProps('shieldCurrent')}
-							{...numberInputProps}
-						/>
-						<NumberInput
-							label='Max Shield'
-							key={form.key('shieldMax')}
-							{...form.getInputProps('shieldMax')}
-							{...numberInputProps}
-						/>
-					</Group>
-					<Group grow align='flex-start'>
-						<NumberInput
-							label='Current Health'
-							key={form.key('healthCurrent')}
-							{...form.getInputProps('healthCurrent')}
-							{...numberInputProps}
-						/>
-						<NumberInput
-							label='Max Health'
-							key={form.key('healthMax')}
-							{...form.getInputProps('healthMax')}
-							{...numberInputProps}
-						/>
-					</Group>
-					<Group>
-						<Button variant='default' onClick={close}>Cancel</Button>
-						<Button flex={1} type='submit'>Update</Button>
-					</Group>
-				</Stack>
-			</form>
+			{/* <form onSubmit={form.onSubmit(handleSubmit)}> */}
+			<Stack>
+				<Group grow>
+					<NumberInput
+						label='Shield Durability'
+						key={form.key('shieldDurability')}
+						{...form.getInputProps('shieldDurability')}
+						{...numberInputProps}
+					/>
+					<NumberInput
+						label='Current Shield'
+						key={form.key('shieldCurrent')}
+						{...form.getInputProps('shieldCurrent')}
+						{...numberInputProps}
+					/>
+					<NumberInput
+						label='Max Shield'
+						key={form.key('shieldMax')}
+						{...form.getInputProps('shieldMax')}
+						{...numberInputProps}
+					/>
+				</Group>
+				<Group grow align='flex-start'>
+					<NumberInput
+						label='Current Health'
+						key={form.key('healthCurrent')}
+						{...form.getInputProps('healthCurrent')}
+						{...numberInputProps}
+					/>
+					<NumberInput
+						label='Max Health'
+						key={form.key('healthMax')}
+						{...form.getInputProps('healthMax')}
+						{...numberInputProps}
+					/>
+				</Group>
+				<Group>
+					<Button variant='default' onClick={close}>Cancel</Button>
+					<Button flex={1} type='submit'>Update</Button>
+				</Group>
+			</Stack>
+			{/* </form> */}
 		</Modal>
 	)
 }
